# swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"MarketPlace Service API","description":"MarketPlace service for products using Async "},"paths":{"/api/Products/allproducts":{"get":{"tags":["Products"],"operationId":"GetProductsAsync","consumes":[],"produces":["text/plain","application/json","text/json","application/xml","text/xml"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Product"}}}}}},"/api/Products/{id}":{"get":{"tags":["Products"],"operationId":"GetProductByIdAsync","consumes":[],"produces":["text/plain","application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Product"}},"404":{"description":"Not Found"}}},"put":{"tags":["Products"],"operationId":"PutProductAsync","consumes":["application/json-patch+json","application/json","text/json","application/*+json","application/xml","text/xml","application/*+xml"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"product","in":"body","required":false,"schema":{"$ref":"#/definitions/Product"}}],"responses":{"204":{"description":"Success"},"400":{"description":"Bad Request"},"404":{"description":"Not Found"}}},"delete":{"tags":["Products"],"operationId":"DeleteAsync","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/api/Products":{"post":{"tags":["Products"],"operationId":"PostProductAsync","consumes":["application/json-patch+json","application/json","text/json","application/*+json","application/xml","text/xml","application/*+xml"],"produces":["text/plain","application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"product","in":"body","required":false,"schema":{"$ref":"#/definitions/Product"}}],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/Product"}},"400":{"description":"Bad Request"}}}}},"definitions":{"Product":{"type":"object","properties":{"productId":{"format":"int32","type":"integer"},"name":{"type":"string"},"price":{"format":"float","type":"number"},"categoryId":{"format":"int32","type":"integer"},"marketId":{"format":"int32","type":"integer"},"discountId":{"format":"int32","type":"integer"},"pictureId":{"type":"string"},"category":{"$ref":"#/definitions/Category"},"discount":{"$ref":"#/definitions/Discount"},"market":{"$ref":"#/definitions/Market"}}},"Category":{"type":"object","properties":{"categoryId":{"format":"int32","type":"integer"},"name":{"type":"string"}}},"Discount":{"type":"object","properties":{"discountId":{"format":"int32","type":"integer"},"offerAsPercent":{"format":"float","type":"number"}}},"Market":{"type":"object","properties":{"marketId":{"format":"int32","type":"integer"},"name":{"type":"string"}}}}}